+++ index.html
@@ -174,7 +174,7 @@
         class="fixed top-1/2 left-1/2 -translate-x-1/2 -translate-y-1/2 z-99999"
       ></div>
 
-      <gutter-ads></gutter-ads>
+      <homepage-promos></homepage-promos>
 
       <!-- Main container with responsive padding -->
       <main-layout class="contents">
