+++ src/client/Main.ts
@@ -27,8 +27,8 @@ import "./GameModeSelector";
 import { GameModeSelector } from "./GameModeSelector";
 import { GameStartingModal } from "./GameStartingModal";
 import "./GoogleAdElement";
-import { GutterAds } from "./GutterAds";
 import { HelpModal } from "./HelpModal";
+import { HomepagePromos } from "./HomepagePromos";
 import { HostLobbyModal as HostPrivateLobbyModal } from "./HostLobbyModal";
 import { JoinLobbyModal } from "./JoinLobbyModal";
 import "./LangSelector";
@@ -60,7 +60,6 @@ import {
 } from "./Utils";
 import "./components/DesktopNavBar";
 import "./components/Footer";
-import "./components/HomeFooterAd";
 import "./components/MainLayout";
 import "./components/MobileNavBar";
 import "./components/PlayPage";
@@ -245,7 +244,6 @@ class Client {
   private tokenLoginModal: TokenLoginModal;
   private matchmakingModal: MatchmakingModal;
 
-  private gutterAds: GutterAds;
   private turnstileTokenPromise: Promise<{
     token: string;
     createdAt: number;
@@ -305,10 +303,9 @@ class Client {
       }
     });
 
-    const gutterAds = document.querySelector("gutter-ads");
-    if (!(gutterAds instanceof GutterAds))
-      throw new Error("Missing gutter-ads");
-    this.gutterAds = gutterAds;
+    const gutterAds = document.querySelector("homepage-promos");
+    if (!(gutterAds instanceof HomepagePromos))
+      throw new Error("Missing homepage-promos");
 
     document.addEventListener("join-lobby", this.handleJoinLobby.bind(this));
     document.addEventListener("leave-lobby", this.handleLeaveLobby.bind(this));
@@ -774,7 +771,7 @@ class Client {
         "token-login",
         "matchmaking-modal",
         "lang-selector",
-        "gutter-ads",
+        "homepage-promos",
       ].forEach((tag) => {
         const modal = document.querySelector(tag) as HTMLElement & {
           close?: () => void;
